Moved utility functions from `assets` to `helpers`
