+++ pages/index.vue
@@ -1248,16 +1248,16 @@
 import url from "url"
 import querystring from "querystring"
 import { commonHeaders } from "~/helpers/headers"
-import parseCurlCommand from "~/assets/js/curlparser.js"
+import parseCurlCommand from "~/helpers/curlparser"
 import getEnvironmentVariablesFromScript from "~/helpers/preRequest"
 import runTestScriptWithVariables from "~/helpers/postwomanTesting"
 import parseTemplateString from "~/helpers/templating"
-import { tokenRequest, oauthRedirect } from "~/assets/js/oauth"
+import { tokenRequest, oauthRedirect } from "~/helpers/oauth"
 import { cancelRunningRequest, sendNetworkRequest } from "~/helpers/network"
 import { fb } from "~/helpers/fb"
 import { getEditorLangForMimeType } from "~/helpers/editorutils"
-import { hasPathParams, addPathParamsToVariables, getQueryParams } from "~/helpers/requestParams.js"
-import { parseUrlAndPath } from "~/helpers/utils/uri.js"
+import { hasPathParams, addPathParamsToVariables, getQueryParams } from "~/helpers/requestParams"
+import { parseUrlAndPath } from "~/helpers/utils/uri"
 import { httpValid } from "~/helpers/utils/valid"
 import { knownContentTypes, isJSONContentType } from "~/helpers/utils/contenttypes"
 import closeIcon from "~/static/icons/close-24px.svg?inline"
